requirements.txt
pyaudio
openai
pvporcupine
boto3
pygame
wave
uuid
gpiozero
sseclient-py
systemd-python
#apt install python3-systemd libsystemd-dev
